requirements.txt
tabulate == 0.8.9
xmltodict==0.12.0
dicttoxml==1.7.4
protobuf==3.15.6
avro==1.10.2
pyyaml==5.4.1
msgpack==1.0.2